deleted methods in employee class, added payslipgenerator class, fixed the equation for getting income tax

+++ PaySlip/PaySlip/TaxRateSchedule.cs
@@ -5,11 +5,11 @@
         public int MinimumTaxableIncome { get; } 
         public int MaximumTaxableIncome { get; } 
         public int NonTaxableIncome { get; } 
-        public int TaxPerDollar { get; } 
+        public double TaxPerDollar { get; } 
         public int ExtraTax { get; } 
         
         
-        public TaxRateSchedule(int minimumTaxableIncome, int maximumTaxableIncome, int nonTaxableIncome, int taxPerDollar, int extraTax)
+        public TaxRateSchedule(int minimumTaxableIncome, int maximumTaxableIncome, int nonTaxableIncome, double taxPerDollar, int extraTax)
         {
             MinimumTaxableIncome = minimumTaxableIncome;
             MaximumTaxableIncome = maximumTaxableIncome;
